+++ src/main/java/com/example/dietitian_plus/domain/unit/dto/UpdateUnitRequestDto.java
@@ -13,6 +13,6 @@ public class UpdateUnitRequestDto {
     @JsonProperty("unit_name")
     private String unitName;
 
-    private Float grams;
+    private Double grams;
 
 }
